+++ R/cvar-package.R
@@ -1,5 +1,6 @@
 #' @author Georgi N. Boshnakov
 #'
+#' @importFrom stats quantile runif
 #' @importFrom gbutils cdf2quantile
 #' @importFrom Rdpack reprompt
 #'
@@ -9,12 +10,12 @@
 #'
 #' @title Compute Conditional Value-at-Risk and Value-at-Risk
 #'
-#' @description Compute expected shortfall (ES) and Value at Risk
-#'     (VaR) from a quantile function, distribution function, random
-#'     number generator or probability density function.  ES is also
-#'     known as Conditional Value at Risk (CVaR). Virtually any
-#'     continuous distribution can be specified.  The functions are
-#'     vectorised over the arguments.
+#' @description Compute expected shortfall (ES) and Value at Risk (VaR)
+#'     from a quantile function, distribution function, random number
+#'     generator or probability density function.  ES is also known as
+#'     Conditional Value at Risk (CVaR). Virtually any continuous
+#'     distribution can be specified.  The functions are vectorised over
+#'     the arguments. Some support for GARCH models is provided, as well.
 #'
 #' @details
 #'
@@ -69,11 +70,15 @@
 #'     transformation, and the possibility to compute it from cdf's
 #'     when quantile functions are not available.
 #'
+#'     Some support for GARCH models is provided, as well. It is
+#'     currently under development, see \code{\link{predict.garch1c1}}
+#'     for current functionality.
+#'
 #' @references
 #'     \insertAllCited{}
 #'
 #' @examples
-#' ## see the examples in for ES() and VaR()
+#' ## see the examples for ES(), VaR(), predict.garch1c1()
 #'
 NULL